% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_sed}
\alias{as_sed}
\title{Format a list of data, read-in, accordingly}
\usage{
as_sed(list, func = "as_sed")
}
\arguments{
\item{list}{The list of dataframes}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\value{
a list of formatted data frames, in the order of same as input.
}
\description{
This function is a substitute for \code{as_sedunit}, allowing to read in a \code{list} of dataframe. The order of the list must be as follow.
\itemize{
\item item 1: the dataframe containing matrix \code{factor}
\item item 2: the dataframe containing matrix \code{item}
\item item 3: the dataframe containing matrix \code{data}
for more \code{item} and \code{data} matrixs, the pattern of \code{item}-\code{data} should continue.
}
}
\examples{
as_sed(list(factor, item1, data1, item2, data2))
}
