% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_FilterRow.R
\name{cal_FilterRow}
\alias{cal_FilterRow}
\title{Filter rows within data of datasets}
\usage{
cal_FilterRow(SED, data = NULL, ...)
}
\arguments{
\item{SED}{SED}

\item{data}{The name of dataset to be converted, in \code{character}. By default (\code{NULL}), all datasets will be converted.}

\item{...}{An expression that returns a logical value, that terms are defined by items in \code{sed_data} of the selected \code{data}. Only rows with an evaluation of \code{TRUE} will be kept. For more information, please check \code{dplyr::filter}, which this function is based on.}
}
\value{
SED
}
\description{
Note that \code{...} expressions must be comprehensible in the dataset. Non-comprehensible expressions will currently return an error.
}
\examples{
cal_FilterRow(SED, data = NULL, a)
}
