% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_group_add.R
\name{cal_group}
\alias{cal_group}
\title{Add grouping to data matrices}
\usage{
cal_group(SED, data = NULL, factor = NULL, overwrite = FALSE)
}
\arguments{
\item{SED}{SED}

\item{data}{The name of dataset to be converted, in \code{character}. By default (\code{NULL}), all datasets will be converted.}

\item{factor}{Factors within a dataset to be added to grouping, in \code{character}. By default (\code{NULL}), all factors will be added to grouping. If the factor is unavailable, said factor will be ignored.}

\item{overwrite}{Should this function overwrite previous grouping? By default (\code{FALSE}), grouping will overwrite previous grouping. For more information, please review the \code{.add} argument in \code{dplyr::group_by}}
}
\value{
SED
}
\description{
Add grouping to data matrices
}
\examples{
cal_group(SED, data = NULL, factor = "#region")
}
