% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_GroupSum.R
\name{cal_GroupSum}
\alias{cal_GroupSum}
\title{Calculate the sum of each group}
\usage{
cal_GroupSum(SED, data = NULL)
}
\arguments{
\item{SED}{The SED}

\item{data}{data}
}
\value{
A SED
}
\description{
Different items with different properties will have different interpretation of \emph{"sum"}.
*Logical: OR operation, where a single TRUE will result in TRUE
*Integer/ Numeric: Numerical add operation
*Character: String conjunction
}
\examples{
cal_GroupSum(SED, data = NULL)
}
