% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ToLogical.R
\name{cal_ToLogical}
\alias{cal_ToLogical}
\title{Convert data to logical data}
\usage{
cal_ToLogical(SED, data = NULL, item)
}
\arguments{
\item{SED}{SED}

\item{data}{The name of dataset to be converted, in \code{character}. By default (\code{NULL}), all datasets will be converted.}

\item{item}{The name of item to be converted across all of \code{data}, in \code{character}. By default (\code{NULL}), all convertible items will be converted into logical data}
}
\value{
SED
}
\description{
Convert data to logical data
}
\examples{
cal_ToLogical(SED, data = "1", item = "a")
}
