% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_group_drop.R
\name{cal_ungroup}
\alias{cal_ungroup}
\title{Drop all grouping from data matrices}
\usage{
cal_ungroup(SED, data = NULL)
}
\arguments{
\item{SED}{SED}

\item{data}{The name of dataset to be converted, in \code{character}. By default (\code{NULL}), all datasets will be converted.}
}
\value{
SED
}
\description{
Drop all grouping from data matrices
}
\examples{
cal_ungroup(SED)
}
