% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataset.R
\name{ck_dataset}
\alias{ck_dataset}
\title{Data: Check data fulfills requirement}
\usage{
ck_dataset(factor, item, data, silent = FALSE, func = "ck_dataset")
}
\arguments{
\item{factor}{the \code{factor} matrix in dataframe or tibble.}

\item{item}{the \code{item} matrix in dataframe or tibble.}

\item{data}{the \code{data} matrix in dataframe or tibble.}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}

\item{func}{The function name that called this function, which is to be passed to \code{sys_msgerror} or \code{sys_msgwarning}. Default as \code{"ck_dataset"}. If \code{NULL}, the function name will not display.}
}
\value{
a logical value, with \code{TRUE} representing an error occurred; and a standardized message if \code{silent} is \code{FALSE}.
}
\description{
Data: Check data fulfills requirement
}
\examples{
ck_dataset(factor, item, data)
}
