% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRW_matrix_data.R
\name{ckrw_data}
\alias{ckrw_data}
\title{Pre-check for r/w: Check the \code{data} matrix if it suits the requirements}
\usage{
ckrw_data(data, silent = FALSE, func = "ckrw_data")
}
\arguments{
\item{data}{the \code{data} matrix in dataframe or tibble.}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}

\item{func}{The function name that called this function, which is to be passed to \code{sys_msgerror} or \code{sys_msgwarning}. Default as \code{"ckrw_data"}. If \code{NULL}, the function name will not display.}
}
\value{
a logical value, with \code{TRUE} representing an error occurred; and a standardized message if \code{silent} is \code{FALSE}.
}
\description{
Pre-check for r/w: Check the \code{data} matrix if it suits the requirements
}
\examples{
ckrw_data(data)
}
