% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRW_sheet.R
\name{ckrw_sheet}
\alias{ckrw_sheet}
\title{Pre-check for r/w: Check the sheet structure of \code{xlsx} file.}
\usage{
ckrw_sheet(file, sheet, silent = F, func = "ckrw_sheet")
}
\arguments{
\item{file}{The file address of the \code{xlsx} file.}

\item{sheet}{The \code{data-sheets} to be read in. By default (\code{NULL}), all \code{data-sheets} will be read in.}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\value{
a logical value, with \code{TRUE} representing an error occurred; and a standardized message if \code{silent} is \code{FALSE}.
}
\description{
Pre-check for r/w: Check the sheet structure of \code{xlsx} file.
}
\examples{
ckrw_sheet(file = DIR)
}
