% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_factor}
\alias{extract_factor}
\title{Extract matrix \code{factor} from an SED}
\usage{
extract_factor(SED, silent = FALSE, func = "extract_factor")
}
\arguments{
\item{SED}{SED object}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}

\item{func}{The function name that called this function, which is to be passed to \code{sys_msg_error} or \code{sys_msgwarning}. Default as \code{NULL}, the function name will not display.}
}
\value{
matrix \code{factor}
}
\description{
Extract matrix \code{factor} from an SED
}
\examples{
extract_factor(SED)
}
