% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sed_set}
\alias{extract_sed_set}
\title{Extract \verb{SED set} from an SED}
\usage{
extract_sed_set(SED, silent = FALSE, func = "extract_factor")
}
\arguments{
\item{SED}{SED object}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}

\item{func}{The function name that called this function, which is to be passed to \code{sys_msg_error} or \code{sys_msgwarning}. Default as \code{NULL}, the function name will not display.}
}
\value{
a list of SED set
}
\description{
Extract \verb{SED set} from an SED
}
\examples{
extract_sed_set(SED)
}
