% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form_read}
\alias{form_read}
\title{Format the data, read-in, accordingly}
\usage{
form_read(factor, item, data, func = "form_read")
}
\arguments{
\item{factor}{The \code{factor} matrix in dataframe or tibble.}

\item{item}{The \code{item} matrix in dataframe or tibble.}

\item{data}{The \code{data} matrix in dataframe or tibble.}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\value{
a list of formatted data frames, in the order of \code{factor}, \code{item}, \code{data}.
}
\description{
Format the data, read-in, accordingly
}
\examples{
form_read(factor, item, data)
}
