% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form_read_multi}
\alias{form_read_multi}
\title{Format a list of data, read-in, accordingly}
\usage{
form_read_multi(list, func = "form_read_multi")
}
\arguments{
\item{list}{The list of dataframes}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\value{
a list of formatted data frames, in the order of same as input.
}
\description{
This function is a substitute for \code{form_read}, allowing to read in a \code{list} of dataframe. The order of the list must be as follow.
\itemize{
\item item 1: the dataframe containing matrix \code{factor}
\item item 2: the dataframe containing matrix \code{item}
\item item 3: the dataframe containing matrix \code{data}
for more \code{item} and \code{data} matrixs, the pattern of \code{item}-\code{data} should continue.
}
}
\examples{
form_read_multi(list(factor, item1, data1, item2, data2))
}
