% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sed}
\alias{is_sed}
\title{Test if the object is a \code{SED}}
\usage{
is_sed(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value. \code{TRUE} for SED data.
}
\description{
Returns \code{TRUE} for SED data, and \code{FALSE} for all other objects, including lists.
}
\examples{
is_sed(x)

}
