% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sed_data}
\alias{is_sed_data}
\title{Test if the object is a \code{SED_data}}
\usage{
is_sed_data(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value. \code{TRUE} for SED data matrix.
}
\description{
Returns \code{TRUE} for SED data matrix, and \code{FALSE} for all other objects, including lists.
}
\examples{
is_sed_data(x)
}
