% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sed_set}
\alias{is_sed_set}
\title{Test if the object is a \code{SED_set}}
\usage{
is_sed_set(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value. \code{TRUE} for SED dataset.
}
\description{
Returns \code{TRUE} for SED dataset, and \code{FALSE} for all other objects, including lists.
}
\examples{
is_sed_set(x)
}
