% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sed}
\alias{print.sed}
\title{Printing SED data}
\usage{
\method{print}{sed}(x)
}
\arguments{
\item{x}{an object}
}
\value{
A visual print of the SED data
}
\description{
Printing SED data
}
\examples{
print(SED)
}
