% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sed.R
\name{read_sed}
\alias{read_sed}
\title{Read a SED file}
\usage{
read_sed(file, sheet, fillMergedCells = TRUE)
}
\arguments{
\item{file}{The file directory itself, with \code{characters}}

\item{sheet}{Data sheets to be read. By default, it will read all the data sheets found.}

\item{fillMergedCells}{Will the funciton fill in all the merged cells in \code{.xlsx} file.}
}
\value{
A SED item
}
\description{
Standard Ecological Data (SED) files are Excel files with a fixed format.
}
\examples{
read_sed(choose.files())
}
