% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_bind.R
\name{sed_bind}
\alias{sed_bind}
\title{Binding 2 SEDs into 1 SED}
\usage{
sed_bind(SED1, SED2, func = "sed_bind", silent = FALSE)
}
\arguments{
\item{SED1}{SED 1}

\item{SED2}{SED 2}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}
}
\value{
a SED
}
\description{
Binding 2 SEDs into 1 SED
}
\examples{
sed_bind(SED1, SED2)
}
