% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_extract.R
\name{sed_extract}
\alias{sed_extract}
\title{Extracting data sets within a SED}
\usage{
sed_extract(SED, data, func = "sed_extract", silent = FALSE)
}
\arguments{
\item{SED}{SED}

\item{data}{Names of the data sets in \code{character}}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}

\item{silent}{Will this function return an error message? Default as \code{FALSE}, which will display a text message.}
}
\value{
a SED
}
\description{
Extracting data sets within a SED
}
\examples{
sed_extract(SED, data = "data_1")
}
