% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_forge.R
\name{sed_forge.character}
\alias{sed_forge.character}
\title{Read a incomplete SED file, and convert it into SED}
\usage{
\method{sed_forge}{character}(x, sheet, flexible_factor = NA, fillMergedCells = TRUE)
}
\arguments{
\item{x}{The file directory itself, with \code{characters}}

\item{sheet}{Data sheets to be read. By default, it will read all the data sheets found.}

\item{flexible_factor}{Name of any flexible factors. By default (\code{NA}), there are no flexible factors.}

\item{fillMergedCells}{Will the function fill in all the merged cells in \code{.xlsx} file? By default (\code{TRUE}), the function will fill merged cell.}
}
\value{
A SED
}
\description{
Read a incomplete SED file, and convert it into SED
}
\examples{
sed_forge.character(choose.files())
}
\keyword{internal}
