% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_forge.R
\name{sed_forge.data.frame}
\alias{sed_forge.data.frame}
\title{Read a dataframe and convert it into SED}
\usage{
\method{sed_forge}{data.frame}(x, flexible_factor = NA)
}
\arguments{
\item{x}{The dataframe itself}

\item{flexible_factor}{Name of any flexible factors. By default (\code{NA}), there are no flexible factors.}
}
\value{
A SED
}
\description{
Read a dataframe and convert it into SED
}
\examples{
sed_forge.data.frame(data)
}
\keyword{internal}
