% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_forge.R
\name{sed_forge}
\alias{sed_forge}
\title{Forge a \code{data} matrix into an SED}
\usage{
sed_forge(x, ...)
}
\arguments{
\item{...}{}

\item{obj}{An object}
}
\value{
A SED
}
\description{
Forge a \code{data} matrix into an SED
}
\examples{
sed_forge(data)
}
