% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_read.R
\name{sed_read}
\alias{sed_read}
\title{Read a xlsx file into a SED object}
\usage{
sed_read(file, sheet, fillMergedCells = TRUE)
}
\arguments{
\item{file}{The file directory itself, with \code{characters}}

\item{sheet}{Data sheets to be read. By default, it will read all the data sheets found.}

\item{fillMergedCells}{Will the function fill in all the merged cells in \code{.xlsx} file? By default (\code{TRUE}), the function will fill merged cell.}
}
\value{
A SED
}
\description{
Standard Ecological Data (SED) files are Excel files with a fixed format.
}
\examples{
sed_read.character(choose.files())
}
