% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_save.R
\name{sed_save}
\alias{sed_save}
\title{Save a SED object into a xlsx file}
\usage{
sed_save(SED, file)
}
\arguments{
\item{SED}{a SED object}

\item{file}{The file directory itself, with \code{characters}}
}
\value{
A xlsx file in the specified directory
}
\description{
Save a SED object into a xlsx file
}
\examples{
sed_save(SED, directory)
}
