% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_msgerror.R
\name{sys_msgerror}
\alias{sys_msgerror}
\title{Create a standardized error message for the package}
\usage{
sys_msgerror(title, error, func = NULL)
}
\arguments{
\item{title}{Title of the error in \code{character}. This will be passed into \code{cli::cli_text}.}

\item{error}{Details of the error in  \code{character}. This will be passed into \code{cli::cli_bullets}. Use \code{"x"} for each bullet.}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\description{
Create a standardized error message for the package
}
\examples{
seal::sys_msgerror("Data incomplete", c("x" = "Please fill in the data"), func = "data_check")
}
