% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_msgwarn.R
\name{sys_msgwarning}
\alias{sys_msgwarning}
\title{Create a standardized error message for the package}
\usage{
sys_msgwarning(title, error, func = NULL)
}
\arguments{
\item{title}{Title of the error in \code{character}. This will be passed into \code{cli::cli_text}.}

\item{error}{Details of the error in  \code{character}. This will be passed into \code{cli::cli_bullets}. Use \code{"!"} for each bullet.}

\item{func}{The function name causing the error in \code{character}. By default \code{NULL}, this will not be displayed.}
}
\description{
Create a standardized error message for the package
}
\examples{
seal::sys_msgwarning("Data incomplete", c("!" = "Please fill in the data"), func = "data_check")
}
